Switch the photontable writer to use h5py for easier installation on windows and embedded platforms
Currently the photontable writer uses pytables which does not package
the native hdf5 library with it. This means it is not easily pip
installable on the board or on platforms where getting build tools is
a pain. In addition the current photontable writter patches pytables
internals which makes it difficult to upgrade pytables. Because
we only need to write photontables here, we can simplify things by
using h5py here.

This change breaks the `photontable` API but it is currently unused
and it is not clear to me that the existing API would work for our
use case in the first place